requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
coverage==7.6.1
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.115.0
flake8==7.1.1
flake8-import-order==0.18.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
psycopg==3.2.3
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-multipart==0.0.12
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
uvicorn==0.31.0
